// src/app/models/stat-collection.ts
export class StatCollection {
    id: string;
    EntType: number;
    EntId: string;
    PA: number;
    AB: number;
    H: number;
    DB: number;
    TB: number;
    HR: number;
    R: number;
    RBI: number;
    SO: number;
    BBFC: number;
    SB: number;
    IP: number;
    ER: number;
    K: number;
    W: number;
    L: number;
    AVG: number;
    OBP: number;
    SLUG: number;
    OPS: number;
    ERA: number;
    K5: number;
}